misc/tests/requirements.txt
py-bcrypt
pymongo
selenium
